+++ NEWS.md
@@ -1,5 +1,10 @@
 # xcms 4.7
 
+## Changes in version
+
+- Fix title of `plotPrecursorIons()` plot.
+
+
 ## Changes in version 4.7.3
 
 - Fix conversion from `XcmsExperiment` objects to `XCMSnExp` (issue #807).